refactor : common 패키지 분리 별도의 컨텍스트에 맞추어 예외처리
