+++ module-product/src/main/java/kitchenpos/product/exception/ProductExceptionHandler.java
@@ -1,0 +1,16 @@
+package kitchenpos.product.exception;
+
+import java.util.NoSuchElementException;
+import org.springframework.http.ResponseEntity;
+import org.springframework.web.bind.annotation.ExceptionHandler;
+import org.springframework.web.bind.annotation.RestControllerAdvice;
+
+@RestControllerAdvice
+public class ProductExceptionHandler {
+
+    @ExceptionHandler({IllegalArgumentException.class, NoSuchElementException.class, IllegalStateException.class})
+    public ResponseEntity<Void> badRequest(RuntimeException e) {
+        return ResponseEntity.badRequest().build();
+    }
+
+}
